Move command line parsing out of fx_muxer.cpp into separate file
