+++ src/corehost/cli/fxr/CMakeLists.txt
@@ -24,6 +24,7 @@ set(SOURCES
     ../fx_definition.cpp
     ../fx_reference.cpp
     ../version.cpp
+    ./command_line.cpp
     ./corehost_init.cpp
     ./hostfxr.cpp
     ./fx_ver.cpp
@@ -49,6 +50,7 @@ set(HEADERS
     ../roll_forward_option.h
     ../roll_fwd_on_no_candidate_fx_option.h
     ../version.h
+    ./command_line.h
     ./corehost_init.h
     ./fx_ver.h
     ./fx_muxer.h
