+++ src/corehost/cli/fxr/command_line.cpp
@@ -1,0 +1,311 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+#include "command_line.h"
+#include <error_codes.h>
+#include "framework_info.h"
+#include <pal.h>
+#include "sdk_info.h"
+#include <trace.h>
+#include <utils.h>
+
+namespace
+{
+    struct host_option
+    {
+        pal::string_t option;
+        pal::string_t argument;
+        pal::string_t description;
+        bool framework_dependent;
+    };
+
+    bool is_sdk_dir_present(const pal::string_t& dotnet_root)
+    {
+        pal::string_t sdk_path = dotnet_root;
+        append_path(&sdk_path, _X("sdk"));
+        return pal::directory_exists(sdk_path);
+    }
+
+    std::vector<host_option> get_known_opts(bool exec_mode, host_mode_t mode, bool for_cli_usage = false)
+    {
+        std::vector<host_option> known_opts = { { _X("--additionalprobingpath"), _X("<path>"), _X("Path containing probing policy and assemblies to probe for.") } };
+        if (for_cli_usage || exec_mode || mode == host_mode_t::split_fx || mode == host_mode_t::apphost)
+        {
+            known_opts.push_back({ _X("--depsfile"), _X("<path>"), _X("Path to <application>.deps.json file.")});
+            known_opts.push_back({ _X("--runtimeconfig"), _X("<path>"), _X("Path to <application>.runtimeconfig.json file.")});
+        }
+
+        if (for_cli_usage || mode == host_mode_t::muxer || mode == host_mode_t::apphost)
+        {
+            // If mode=host_mode_t::apphost, these are only used when the app is framework-dependent.
+            known_opts.push_back({ _X("--fx-version"), _X("<version>"), _X("Version of the installed Shared Framework to use to run the application.") });
+            known_opts.push_back({ _X("--roll-forward"), _X("<value>"), _X("Roll forward to framework version (LatestPatch, Minor, LatestMinor, Major, LatestMajor, Disable)") });
+            known_opts.push_back({ _X("--additional-deps"), _X("<path>"), _X("Path to additional deps.json file.") });
+
+            if (!for_cli_usage)
+            {
+                // Intentionally leave this one out of for_cli_usage since we don't want to show it in command line help (it's deprecated).
+                known_opts.push_back({ _X("--roll-forward-on-no-candidate-fx"), _X("<n>"), _X("<obsolete>") });
+            }
+        }
+
+        return known_opts;
+    }
+
+    bool parse_known_args(
+        const int argc,
+        const pal::char_t* argv[],
+        const std::vector<host_option>& known_opts,
+        // Although multimap would provide this functionality the order of kv, values are
+        // not preserved in C++ < C++0x
+        opt_map_t* opts,
+        int* num_args)
+    {
+        int arg_i = *num_args;
+        while (arg_i < argc)
+        {
+            pal::string_t arg = argv[arg_i];
+            pal::string_t arg_lower = pal::to_lower(arg);
+            if (std::find_if(known_opts.begin(), known_opts.end(),
+                [&](const host_option& hostoption) { return arg_lower == hostoption.option; })
+                == known_opts.end())
+            {
+                // Unknown argument.
+                break;
+            }
+
+            // Known argument, so expect one more arg (value) to be present.
+            if (arg_i + 1 >= argc)
+            {
+                return false;
+            }
+
+            trace::verbose(_X("Parsed known arg %s = %s"), arg.c_str(), argv[arg_i + 1]);
+            (*opts)[arg_lower].push_back(argv[arg_i + 1]);
+
+            // Increment for both the option and its value.
+            arg_i += 2;
+        }
+
+        *num_args = arg_i;
+
+        return true;
+    }
+
+    int parse_args(
+        const host_startup_info_t &host_info,
+        int argoff,
+        int argc,
+        const pal::char_t *argv[],
+        bool exec_mode,
+        host_mode_t mode,
+        /*out*/ int *new_argoff,
+        /*out*/ pal::string_t &app_candidate,
+        /*out*/ opt_map_t &opts)
+    {
+        std::vector<host_option> known_opts = get_known_opts(exec_mode, mode);
+
+        // Parse the known arguments if any.
+        int num_parsed = 0;
+        if (!parse_known_args(argc - argoff, &argv[argoff], known_opts, &opts, &num_parsed))
+        {
+            trace::error(_X("Failed to parse supported options or their values:"));
+            for (const auto& arg : known_opts)
+            {
+                trace::error(_X("  %-37s  %s"), (arg.option + _X(" ") + arg.argument).c_str(), arg.description.c_str());
+            }
+            return StatusCode::InvalidArgFailure;
+        }
+
+        app_candidate = host_info.app_path;
+        *new_argoff = argoff + num_parsed;
+        bool doesAppExist = false;
+        if (mode == host_mode_t::apphost)
+        {
+            doesAppExist = pal::realpath(&app_candidate);
+        }
+        else
+        {
+            trace::verbose(_X("Using the provided arguments to determine the application to execute."));
+            if (*new_argoff >= argc)
+            {
+                command_line::print_muxer_usage(!is_sdk_dir_present(host_info.dotnet_root));
+                return StatusCode::InvalidArgFailure;
+            }
+
+            app_candidate = argv[*new_argoff];
+
+            bool is_app_managed = ends_with(app_candidate, _X(".dll"), false) || ends_with(app_candidate, _X(".exe"), false);
+            if (!is_app_managed)
+            {
+                trace::verbose(_X("Application '%s' is not a managed executable."), app_candidate.c_str());
+                if (!exec_mode)
+                {
+                    // Route to CLI.
+                    return StatusCode::AppArgNotRunnable;
+                }
+            }
+
+            doesAppExist = pal::realpath(&app_candidate);
+            if (!doesAppExist)
+            {
+                trace::verbose(_X("Application '%s' does not exist."), app_candidate.c_str());
+                if (!exec_mode)
+                {
+                    // Route to CLI.
+                    return StatusCode::AppArgNotRunnable;
+                }
+            }
+
+            if (!is_app_managed && doesAppExist)
+            {
+                assert(exec_mode == true);
+                trace::error(_X("dotnet exec needs a managed .dll or .exe extension. The application specified was '%s'"), app_candidate.c_str());
+                return StatusCode::InvalidArgFailure;
+            }
+        }
+
+        // App is managed executable.
+        if (!doesAppExist)
+        {
+            trace::error(_X("The application to execute does not exist: '%s'"), app_candidate.c_str());
+            return StatusCode::InvalidArgFailure;
+        }
+
+        return StatusCode::Success;
+    }
+}
+
+pal::string_t command_line::get_last_known_arg(
+    const opt_map_t& opts,
+    const pal::string_t& opt_key,
+    const pal::string_t& de_fault)
+{
+    if (opts.count(opt_key))
+    {
+        const auto& val = opts.find(opt_key)->second;
+        return val[val.size() - 1];
+    }
+    return de_fault;
+}
+
+int command_line::parse_args_for_mode(
+    host_mode_t mode,
+    const host_startup_info_t &host_info,
+    const int argc,
+    const pal::char_t *argv[],
+    /*out*/ int *new_argoff,
+    /*out*/ pal::string_t &app_candidate,
+    /*out*/ opt_map_t &opts)
+{
+    int result;
+    if (mode == host_mode_t::split_fx)
+    {
+        // Invoked as corehost
+        trace::verbose(_X("--- Executing in split/FX mode..."));
+        result = parse_args(host_info, 1, argc, argv, false, mode, new_argoff, app_candidate, opts);
+    }
+    else if (mode == host_mode_t::apphost)
+    {
+        // Invoked from the application base.
+        trace::verbose(_X("--- Executing in a native executable mode..."));
+        result = parse_args(host_info, 1, argc, argv, false, mode, new_argoff, app_candidate, opts);
+    }
+    else
+    {
+        // Invoked as the dotnet.exe muxer.
+        assert(mode == host_mode_t::muxer);
+        trace::verbose(_X("--- Executing in muxer mode..."));
+
+        if (argc <= 1)
+        {
+            command_line::print_muxer_usage(!is_sdk_dir_present(host_info.dotnet_root));
+            return StatusCode::InvalidArgFailure;
+        }
+
+        if (pal::strcasecmp(_X("exec"), argv[1]) == 0)
+        {
+            // arg offset 2 for dotnet, exec
+            result = parse_args(host_info, 2, argc, argv, true, mode, new_argoff, app_candidate, opts);
+        }
+        else
+        {
+            // arg offset 1 for dotnet
+            result = parse_args(host_info, 1, argc, argv, false, mode, new_argoff, app_candidate, opts);
+        }
+    }
+
+    return result;
+}
+
+int command_line::parse_args_for_sdk_command(
+    const host_startup_info_t& host_info,
+    const int argc,
+    const pal::char_t* argv[],
+    /*out*/ int *new_argoff,
+    /*out*/ pal::string_t &app_candidate,
+    /*out*/ opt_map_t &opts)
+{
+    // arg offset 1 for dotnet
+    return parse_args(host_info, 1, argc, argv, false, host_mode_t::muxer, new_argoff, app_candidate, opts);
+}
+
+void command_line::print_muxer_info(const pal::string_t &dotnet_root)
+{
+    trace::println();
+    trace::println(_X("Host (useful for support):"));
+    trace::println(_X("  Version: %s"), _STRINGIFY(HOST_FXR_PKG_VER));
+
+    pal::string_t commit = _STRINGIFY(REPO_COMMIT_HASH);
+    trace::println(_X("  Commit:  %s"), commit.substr(0, 10).c_str());
+
+    trace::println();
+    trace::println(_X(".NET Core SDKs installed:"));
+    if (!sdk_info::print_all_sdks(dotnet_root, _X("  ")))
+    {
+        trace::println(_X("  No SDKs were found."));
+    }
+
+    trace::println();
+    trace::println(_X(".NET Core runtimes installed:"));
+    if (!framework_info::print_all_frameworks(dotnet_root, _X("  ")))
+    {
+        trace::println(_X("  No runtimes were found."));
+    }
+
+    trace::println();
+    trace::println(_X("To install additional .NET Core runtimes or SDKs:"));
+    trace::println(_X("  %s"), DOTNET_CORE_DOWNLOAD_URL);
+}
+
+void command_line::print_muxer_usage(bool is_sdk_present)
+{
+    std::vector<host_option> known_opts = get_known_opts(true, host_mode_t::invalid, /*for_cli_usage*/ true);
+
+    if (!is_sdk_present)
+    {
+        trace::println();
+        trace::println(_X("Usage: dotnet [host-options] [path-to-application]"));
+        trace::println();
+        trace::println(_X("path-to-application:"));
+        trace::println(_X("  The path to an application .dll file to execute."));
+    }
+    trace::println();
+    trace::println(_X("host-options:"));
+
+    for (const auto& arg : known_opts)
+    {
+        trace::println(_X("  %-30s  %s"), (arg.option + _X(" ") + arg.argument).c_str(), arg.description.c_str());
+    }
+    trace::println(_X("  --list-runtimes                 Display the installed runtimes"));
+    trace::println(_X("  --list-sdks                     Display the installed SDKs"));
+
+    if (!is_sdk_present)
+    {
+        trace::println();
+        trace::println(_X("Common Options:"));
+        trace::println(_X("  -h|--help                           Displays this help."));
+        trace::println(_X("  --info                              Display .NET Core information."));
+    }
+}
