+++ src/corehost/common/utils.cpp
@@ -200,59 +200,6 @@ const pal::char_t* get_arch()
 #endif
 }
 
-pal::string_t get_last_known_arg(
-    const opt_map_t& opts,
-    const pal::string_t& opt_key,
-    const pal::string_t& de_fault)
-{
-    if (opts.count(opt_key))
-    {
-        const auto& val = opts.find(opt_key)->second;
-        return val[val.size() - 1];
-    }
-    return de_fault;
-}
-
-bool parse_known_args(
-    const int argc,
-    const pal::char_t* argv[],
-    const std::vector<host_option>& known_opts,
-    // Although multimap would provide this functionality the order of kv, values are
-    // not preserved in C++ < C++0x
-    opt_map_t* opts,
-    int* num_args)
-{
-    int arg_i = *num_args;
-    while (arg_i < argc)
-    {
-        pal::string_t arg = argv[arg_i];
-        pal::string_t arg_lower = pal::to_lower(arg);
-        if (std::find_if(known_opts.begin(), known_opts.end(),
-            [&](const host_option& hostoption) { return arg_lower == hostoption.option; })
-            == known_opts.end())
-        {
-            // Unknown argument.
-            break;
-        }
-
-        // Known argument, so expect one more arg (value) to be present.
-        if (arg_i + 1 >= argc)
-        {
-            return false;
-        }
-
-        trace::verbose(_X("Parsed known arg %s = %s"), arg.c_str(), argv[arg_i + 1]);
-        (*opts)[arg_lower].push_back(argv[arg_i + 1]);
-
-        // Increment for both the option and its value.
-        arg_i += 2;
-    }
-
-    *num_args = arg_i;
-
-    return true;
-}
-
 // Try to match 0xEF 0xBB 0xBF byte sequence (no endianness here.)
 bool skip_utf8_bom(pal::istream_t* stream)
 {
