+++ src/corehost/common/utils.h
@@ -7,13 +7,6 @@
 
 #include "pal.h"
 #include "trace.h"
-struct host_option
-{
-    pal::string_t option;
-    pal::string_t argument;
-    pal::string_t description;
-    bool framework_dependent;
-};
 
 #define _STRINGIFY(s) _X(s)
 #if defined(_WIN32)
@@ -27,8 +20,6 @@ struct host_option
 #define DOTNET_CORE_DOWNLOAD_URL _X("https://aka.ms/dotnet-download")
 
 #define RUNTIME_STORE_DIRECTORY_NAME _X("store")
-
-typedef std::unordered_map<pal::string_t, std::vector<pal::string_t>> opt_map_t;
 
 bool ends_with(const pal::string_t& value, const pal::string_t& suffix, bool match_case);
 bool starts_with(const pal::string_t& value, const pal::string_t& prefix, bool match_case);
@@ -44,16 +35,6 @@ void remove_trailing_dir_seperator(pal::string_t* dir);
 void replace_char(pal::string_t* path, pal::char_t match, pal::char_t repl);
 pal::string_t get_replaced_char(const pal::string_t& path, pal::char_t match, pal::char_t repl);
 const pal::char_t* get_arch();
-pal::string_t get_last_known_arg(
-    const opt_map_t& opts,
-    const pal::string_t& opt_key,
-    const pal::string_t& de_fault);
-bool parse_known_args(
-    const int argc,
-    const pal::char_t* argv[],
-    const std::vector<host_option>& known_opts,
-    opt_map_t* opts,
-    int* num_args);
 bool skip_utf8_bom(pal::istream_t* stream);
 bool get_env_shared_store_dirs(std::vector<pal::string_t>* dirs, const pal::string_t& arch, const pal::string_t& tfm);
 bool get_global_shared_store_dirs(std::vector<pal::string_t>* dirs, const pal::string_t& arch, const pal::string_t& tfm);
